+++ src/lib/components/footer/footer.svelte
