+++ src/lib/components/speakers/speakersList.svelte
