+++ src/lib/components/videos/videoThumbnail.svelte
