+++ src/routes/(subPages)/speakers/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import SpeakersList from '$lib/components/speakersList.svelte';
+  import SpeakersList from '$lib/components/speakers/speakersList.svelte';
 </script>
 
 <h1>Speakers</h1>
