+++ src/routes/(subPages)/videos/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import VideoList from '$lib/components/videoList.svelte';
+  import VideoList from '$lib/components/videos/videoList.svelte';
 </script>
 
 <h1>Videos</h1>
